Move keyboard dom node nav util
